arsync::i18n

Enum TranslationKey

pub enum TranslationKey {

    ProgressDiscovered,
    ProgressCompleted,
    ProgressInFlight,
    ProgressComplete,
    ProgressFiles,
    ProgressBytes,
    ProgressSpeed,
    ProgressEta,
    StatusCopyingFile,
    StatusCopyingDirectory,
    StatusCreatingSymlink,
    StatusCreatingHardlink,
    StatusPreservingMetadata,
    StatusComplete,
    StatusFailed,
    ErrorFileNotFound,
    ErrorPermissionDenied,
    ErrorIoError,
    ErrorInvalidPath,
    ErrorSourceNotExists,
    ErrorDestinationExists,
    InfoStartingCopy,
    InfoScanningSource,
    InfoCreatingDestination,
    InfoPreservingPermissions,
    InfoPreservingOwnership,
    InfoPreservingTimestamps,
    InfoPreservingXattrs,
    InfoPreservingAcls,
    InfoDryRun,
    HelpDescription,
    HelpSource,
    HelpDestination,
    HelpArchive,
    HelpRecursive,
    HelpLinks,
    HelpPerms,
    HelpTimes,
    HelpGroup,
    HelpOwner,
    HelpDevices,
    HelpXattrs,
    HelpAcls,
    HelpHardLinks,
    HelpVerbose,
    HelpQuiet,
    HelpProgress,
    HelpDryRun,
    HelpPirate,
    HelpQueueDepth,
    HelpMaxFilesInFlight,
    HelpCpuCount,
    HelpBufferSize,
    UnitBytes,
    UnitKilobytes,
    UnitMegabytes,
    UnitGigabytes,
    UnitSeconds,
    MiscIn,
    MiscOf,
    MiscAverage,
    MiscTotal,

}

Translation key enum - maps to Fluent message IDs

Variants

ProgressDiscovered

ProgressCompleted

ProgressInFlight

ProgressComplete

ProgressFiles

ProgressBytes

ProgressSpeed

ProgressEta

StatusCopyingFile

StatusCopyingDirectory

StatusCreatingSymlink

StatusCreatingHardlink

StatusPreservingMetadata

StatusComplete

StatusFailed

ErrorFileNotFound

ErrorPermissionDenied

ErrorIoError

ErrorInvalidPath

ErrorSourceNotExists

ErrorDestinationExists

InfoStartingCopy

InfoScanningSource

InfoCreatingDestination

InfoPreservingPermissions

InfoPreservingOwnership

InfoPreservingTimestamps

InfoPreservingXattrs

InfoPreservingAcls

InfoDryRun

HelpDescription

HelpSource

HelpDestination

HelpArchive

HelpRecursive

HelpLinks

HelpPerms

HelpTimes

HelpGroup

HelpOwner

HelpDevices

HelpXattrs

HelpAcls

HelpHardLinks

HelpVerbose

HelpQuiet

HelpProgress

HelpDryRun

HelpPirate

HelpQueueDepth

HelpMaxFilesInFlight

HelpCpuCount

HelpBufferSize

UnitBytes

UnitKilobytes

UnitMegabytes

UnitGigabytes

UnitSeconds

MiscIn

MiscOf

MiscAverage

MiscTotal

Implementations

impl TranslationKey

const fn message_id(self) -> &'static str

Get the Fluent message ID for this key

pub fn get(self) -> Result<String, I18nError>

Get the translated string for this key in the current language

Errors

Returns I18nError::LockPoisoned if the locale lock is poisoned

pub fn translate(self, locale: &str) -> String

Get the translated string for this key in a specific language