arsync::protocol::ssh

Struct SshConnection

pub struct SshConnection {
    process: Child,
    stdin: ChildStdin,
    stdout: ChildStdout,
    host: String,
    user: String,
}

SSH connection to remote host

Uses compio::process for async process management with io_uring backend.

Fields
process: Child

SSH process

stdin: ChildStdin

stdin pipe to remote arsync

stdout: ChildStdout

stdout pipe from remote arsync

host: String

Remote host

user: String

Remote user

Implementations

impl SshConnection

pub async fn connect(host: &str, user: &str, remote_shell: &str) -> Result<Self>

Connect to remote host via SSH

Spawns an SSH process connecting to the remote host and starting arsync in server mode.

Arguments

host - Remote hostname or IP

user - Remote username

remote_shell - Shell command to use (typically “ssh”)

Errors

Returns an error if:

SSH process fails to spawn

Cannot get stdin/stdout from process

Example

use arsync::protocol::ssh::SshConnection;

let conn = SshConnection::connect("example.com", "user", "ssh").await?;

pub async fn start_server(&mut self, _path: &Path) -> Result<()>

Start remote server (send initial protocol negotiation)

Errors

Currently always succeeds. Will return errors in future when protocol negotiation is implemented.